Update patch set 3
Patch Set 3:

I refactored the code so as to increase the code coverage(I kept some classes for reference during plugin development.But now I had to remove them). And added one test class for the util method.
But I see no difference in the sonar report although in local the coverage report is different than what is shown in the logs.

Patch-set: 3